fixing liked movies architecture
