+++ MovieFinderAPI/DtoModels/LikedMoviesDto.cs
@@ -5,5 +5,7 @@
         public int LikedId { get; set; }
         public int MovieId { get; set; }
         public int UserId { get; set; }
+        public string Title { get; set; }
+        public string Poster { get; set; }
     }
 }
