+++ MovieFinderAPI/Models/LikedMovies.cs
@@ -7,5 +7,7 @@ namespace MovieFinder.Models
         public int LikedId { get; set; }
         public int MovieId { get; set; }
         public int UserId { get; set; }
+        public string Title { get; set; }
+        public string Poster { get; set; }
     }
 }
