+++ MovieFinderAPI/Repository/Interface/ILikedMoviesRepository.cs
@@ -5,6 +5,6 @@ namespace MovieFinder.Repository.Interface
 {
     public interface ILikedMoviesRepository : IMovieFinderRepository<LikedMovies>
     {
-        IEnumerable<LikedMovies> GetAllByUserId(int userId, int? skip, int? count);
+        IEnumerable<LikedMovies> GetAllByUserId(int userId, int? skip = null, int? count = null);
     }
 }
